Expose packCircles, packEnclosingCircle.
