+++ index.js
@@ -1,6 +1,8 @@
 export {version} from "./build/package";
 export {default as hierarchy} from "./src/hierarchy";
 export {default as pack} from "./src/pack/index";
+export {default as packCircles} from "./src/pack/circles";
+export {default as packEnclosingCircle} from "./src/pack/enclosingCircle";
 export {default as partition} from "./src/partition";
 export {default as treemap} from "./src/treemap/index";
 export {default as treemapBinary} from "./src/treemap/binary";
