--- src/pack/intersects.js
@@ -1,6 +1,0 @@
-export default function(a, b) {
-  var dx = b.x - a.x,
-      dy = b.y - a.y,
-      dr = a.r + b.r;
-  return dr * dr > dx * dx + dy * dy;
-}
